+++ after/plugin/nvim-tmux-navigation.lua
@@ -1,0 +1,12 @@
+require("nvim-tmux-navigation").setup {
+  disable_when_zoomed = true,
+
+  keybindings = {
+    down = "<C-j>",
+    last_active = "<C-\\>",
+    left = "<C-h>",
+    next = "<C-Space>",
+    right = "<C-l>",
+    up = "<C-k>",
+  },
+}